Use a more succinct method to run commands
Sometimes it's useful to take a moment to read the docs:
https://docs.python.org/3/library/subprocess.html#subprocess.run

This doesn't imply we did anything wrong; making the test pass is
valuable because it allowed me to make this change with confidence.

Given that our use case doesn't really need to parse any output, nor
provide any input, the .run() method is easier to use.

Note that it still uses .Popen() behind the scenes. We can use the more
powerful interface when our needs warrant it.

The other change is to be less specific on the error code. What we
"know" is that `ls` will return `0` (aka success) if the specified
file/directory exists. We really don't care what specific error code it
returns.