Actually check the subprocess exit code
Funny how the previous code actually worked! Comparing a
subprocess.CompletedProcess to an integer doesn't result in a runtime
error so the test was accidentally working.

Again, taking a moment to read the docs:
https://docs.python.org/3/library/subprocess.html#subprocess.CompletedProcess

One can see that the object has a `returncode` property that can be
inspected. Notice how the docs reads "Exit status of the child
process.", this is a good example how terminology in one system (unix
processes) and gets renamed for the python system. Exit Code is a relic
of the past and Return Code is a more accurate name. There are some
tools that can exit non-zero and still be considered a successful
execution of the process.