Add HEAD check to empty repository scenario
This is an important test because this is one of the first things that's
explained when facilitating the workshop.

Let's review the "And .git/refs contains a directory named heads" and
see if there are any changes we would like to make to it.

Bonus question... why is .strip() necessary in the step that I added? Is
there another way I could have written the assert to make that not
necessary?