Always specify main as the initial branch
On one of my machines with a different git configuration, it warned
about using master in some of the commands.

```
hint: Using 'master' as the name for the initial branch. This default branch name
hint: is subject to change. To configure the initial branch name to use in all
hint: of your new repositories, which will suppress this warning, call:
hint:
hint: 	git config --global init.defaultBranch <name>
hint:
hint: Names commonly chosen instead of 'master' are 'main', 'trunk' and
hint: 'development'. The just-created branch can be renamed via this command:
hint:
```

Given that the initial branch override isn't used, it seemed best to
hard code to main for now.