Assert that creating a branch does'nt change HEAD
So cool that I could add a step and not have to implement any step
definition code! There's quite a library of steps emerging now :-)

Maybe there's some adjusting to the wording we can do; but for now it's
this shows the power of creating a library. The wonders of re-use!