Add style linting
This is being added because I wanted to make a structure change to some
code but wasn't 100% sure if it is concerned "standard" python.

Linting is awesome when you're not an expert in a language. Experts can
break the rules because they know why the rule exist.

This commit will now fail the tests if the code doesn't follow standard
pep8 style.

Is this a bit strict? Yes! But better to start with standardized code
first and break the rules later.

The next commit will have all the fixes, but try running the tests at
this commit and see if the reports make sense.