Speculating that the git commits need an author
* Also allow the gha workflow file to trigger the build so that I don't
  need to make arbitrary commits in git-workshop.
* Sorting the env to make it easier to eye-ball.

Git will complain if there isn't any author configuration. That's the
minimum it needs to perform a commit. This might be useful to bake
into the tests, because when running the tests, it'll take the users git
configuration for making commits. This is on reason the workshop runs in
a docker image; to isolate user configuration.